camera: scan a cropped section of the full camera resolution
The insight driving this change are two-fold:

First, that the RPi first generation camera has a fixed focus point
about 50cm from the lens, which means that the resolution gained from
bringing the QR code closer is (more than) lost by the out-of-focus blur.

Second, for performance reasons we're not scanning the full resoluton
of the sensor (2592x1944), but rather a scaled down image delivered
by the scaler hardware.

So, it's much better to configure the sensor for full resolution and
scan a cropped center section of that, in other words scanning a digital
zoom instead of a scaled one. The result is higher resolution at
the same processing cost, and s a bonus the zoomed viewfinder image
naturally leads the user to move the QR codes further away, sharpening
the scan.

The bulk of the change is switching from the V4L2 compatibility API to
the libcamera2 C++ API, because only the C++ API is flexible enough to
express full resolution digital zoom.