lcd,rgb16,gui: use the panel-mipi-dbi driver for the Waveshare display
The Linux kernel can drive the display faster than from userspace.

There are two drivers, one for DRM and one for the older framebuffer
interface. The framebuffer interface is simpler but refreshes continuously
leading to incomplete frames. The DRM API offers an interface for driving
the updates. There is still no vsync, but that's a hardware limitation.