flake.nix: statically include libstdc++, libgcc_s, libatomic
For some reason, libstdc++.so slightly differs according to build platform,
thus ruining reproducibility. Fortunately, the difference disappears if
building with the flag `-static-libstdc++`, at the cost of slightly
larger binaries.

While we're at it, also build with `-static-libgcc` to statically embed
libgcc_s.so and libatomic.so.

In total, we regain reproducibility and simplify the build image a bit.