bc/urtypes: add support for engraving single-sig crypto-accounts
The UR data type crypto-account is used by signers such as SeedSigner
to export xpubs. For multi-sig the crypto-account is intended to be
imported in a coordinator such as Sparrow to participate in a multi-sig
wallet. However, single-sig crypto-accounts contain complete output
descriptors and as such are ready to be engraved.

This change implements limited support for crypto-account, by treating
them as wrapped crypto-outputs (output descriptor), allowing backup of
single-sig wallets directly from the signer.

While here, remove a debug vestige.