gui: remove seed-only backup mode
Before this change, there were two options on the main selection screen:
singlesig and multisig. The names were confusing, because "singlesig"
really meant "seed only" and "multisig" meant "with descriptor".

This change gets rid of the descriptor-less "singlesig" mode and renames
the remaining "multisig" mode to just "backup wallet".

With this change, there is no longer a way to engrave only seeds on
plates. We believe this is an improvement: the interface has fewer
options, and every backup will now include a descriptor that precisely
describes the wallet.